+++ otx/algorithms/detection/adapters/mmdet/task.py
@@ -60,8 +60,8 @@ from otx.algorithms.detection.adapters.mmdet.configurer import (
     SemiSLDetectionConfigurer,
 )
 from otx.algorithms.detection.adapters.mmdet.datasets import ImageTilingDataset
-from otx.algorithms.detection.adapters.mmdet.hooks.det_saliency_map_hook import (
-    DetSaliencyMapHook,
+from otx.algorithms.detection.adapters.mmdet.hooks.det_class_probability_map_hook import (
+    DetClassProbabilityMapHook,
 )
 from otx.algorithms.detection.adapters.mmdet.utils import patch_tiling
 from otx.algorithms.detection.adapters.mmdet.utils.builder import build_detector
@@ -74,6 +74,7 @@ from otx.algorithms.detection.utils import get_det_model_api_configuration
 from otx.api.configuration import cfg_helper
 from otx.api.configuration.helper.utils import config_to_bytes, ids_to_strings
 from otx.api.entities.datasets import DatasetEntity
+from otx.api.entities.explain_parameters import ExplainParameters
 from otx.api.entities.inference_parameters import InferenceParameters
 from otx.api.entities.model import (
     ModelEntity,
@@ -385,8 +386,8 @@ class MMDetectionTask(OTXDetectionTask):
         mm_dataset = build_dataset(cfg.data.test)
         dataloader = build_dataloader(
             mm_dataset,
-            samples_per_gpu=cfg.data.get("samples_per_gpu", 1),
-            workers_per_gpu=cfg.data.get("workers_per_gpu", 0),
+            samples_per_gpu=samples_per_gpu,
+            workers_per_gpu=cfg.data.test_dataloader.get("workers_per_gpu", 0),
             num_gpus=len(cfg.gpu_ids),
             dist=cfg.distributed,
             seed=cfg.get("seed", None),
@@ -438,7 +439,7 @@ class MMDetectionTask(OTXDetectionTask):
             if isinstance(raw_model, TwoStageDetector):
                 saliency_hook = ActivationMapHook(feature_model)
             else:
-                saliency_hook = DetSaliencyMapHook(feature_model)
+                saliency_hook = DetClassProbabilityMapHook(feature_model)
 
         if not dump_features:
             feature_vector_hook: Union[nullcontext, BaseRecordingForwardHook] = nullcontext()
@@ -591,6 +592,7 @@ class MMDetectionTask(OTXDetectionTask):
         output_model.set_data("config.json", config_to_bytes(self._hyperparams))
         output_model.precision = self._precision
         output_model.optimization_methods = self._optimization_methods
+        output_model.has_xai = dump_features
         output_model.set_data(
             "label_schema.json",
             label_schema_to_bytes(self._task_environment.label_schema),
@@ -600,12 +602,12 @@ class MMDetectionTask(OTXDetectionTask):
     def explain(
         self,
         dataset: DatasetEntity,
-        explain_parameters: Optional[InferenceParameters] = None,
+        explain_parameters: Optional[ExplainParameters] = None,
     ) -> DatasetEntity:
         """Main explain function of MMDetectionTask."""
 
         explainer_hook_selector = {
-            "classwisesaliencymap": DetSaliencyMapHook,
+            "classwisesaliencymap": DetClassProbabilityMapHook,
             "eigencam": EigenCamHook,
             "activationmap": ActivationMapHook,
         }