+++ otx/algorithms/segmentation/tasks/inference.py
@@ -89,7 +89,7 @@ class SegmentationInferenceTask(BaseTask, IInferenceTask, IExportTask, IEvaluati
         # self._should_stop = False
         self.freeze = True
         self.metric = "mDice"
-        self._label_dictionary = {}  # type: Dict
+        self._label_dictionary: Dict = {}
 
         super().__init__(SegmentationConfig, task_environment, **kwargs)
         self._label_dictionary = dict(enumerate(sorted(self._labels), 1))
@@ -183,6 +183,7 @@ class SegmentationInferenceTask(BaseTask, IInferenceTask, IExportTask, IEvaluati
             output_model.set_data("openvino.xml", f.read())
         output_model.precision = self._precision
         output_model.optimization_methods = self._optimization_methods
+        output_model.has_xai = dump_features
         output_model.set_data("label_schema.json", label_schema_to_bytes(self._task_environment.label_schema))
         logger.info("Exporting completed")