+++ otx/core/ov/ops/op.py
@@ -46,7 +46,7 @@ class Operation(torch.nn.Module, Generic[_T]):  # pylint: disable=abstract-metho
         op_type = ov_op.get_type_name()
         op_version = ov_op.get_version()
         op_name = get_op_name(ov_op)
-        assert cls.TYPE != "" and cls.VERSION >= 0
+        assert cls.TYPE and cls.VERSION >= 0
         assert op_type == cls.TYPE
         assert op_version == cls.VERSION